backend/pyproject.toml
[project]
name = "tavolaperfetta-backend"
version = "0.1.0"
description = "Tavola Perfetta backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy==2.0.35",
    "psycopg2-binary==2.9.9",
    "alembic==1.13.2",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "python-multipart==0.0.9",
    "google-generativeai==0.8.1",
    "python-dotenv==1.0.1",
    "pyjwt==2.9.0",
    "httpx==0.27.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "faker>=24.0.0",
    "factory-boy>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["app"]
branch = true
omit = ["app/__pycache__/*", "tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
show_missing = true
